fix Massing namespaces & build settings
- move Building to Generate namespace and types to sub-namespace
- replace relative paths to RevitAPI with absolute
