+++ src/MassingSandbox/Generate/BuildingBase.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading;
 using Autodesk.DesignScript.Geometry;
 
-namespace GenerativeToolkit
+namespace Autodesk.RefineryToolkits.Massing.Generate
 {
     internal abstract class BuildingBase : IDisposable
     {
